+++ carp_mobile_sensing/example/lib/example.dart
@@ -1,7 +1,7 @@
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 
 /// This is the code for the very minimal example used in the README.md file.
-void example() {
+void example() async {
   // Create a study using a File Backend
   Study study = Study("1234", "bardram", name: "bardram study");
   study.dataEndPoint = FileDataEndPoint()
@@ -9,17 +9,9 @@ void example() {
     ..zip = true
     ..encrypt = false;
 
-  // add a task to collect location, activity, and weather information
-  study.addTask(SequentialTask('Location, Activity, and Weather Task')
-    ..addMeasure(Measure(MeasureType(NameSpace.CARP, DataType.LOCATION)))
-    ..addMeasure(Measure(MeasureType(NameSpace.CARP, DataType.ACTIVITY))));
-//    ..addMeasure(WeatherMeasure(MeasureType(NameSpace.CARP, DataType.WEATHER))
-//      ..enabled = true
-//      ..frequency = 2 * 60 * 60 * 1000));
-
   // add sensor collection from accelerometer and gyroscope
   // careful - these sensors generate a lot of data!
-  study.addTask(ParallelTask('Sensor Task')
+  study.addTask(Task('Sensor Task')
     ..addMeasure(PeriodicMeasure(MeasureType(NameSpace.CARP, DataType.ACCELEROMETER),
         frequency: 10 * 1000, // sample every 10 secs)
         duration: 100 // for 100 ms
@@ -29,43 +21,19 @@ void example() {
         duration: 100 // for 100 ms
         )));
 
-  study.addTask(Task('Audio Recording Task')
-    ..addMeasure(AudioMeasure(MeasureType(NameSpace.CARP, DataType.AUDIO),
-        frequency: 10 * 60 * 1000, // sample sound every 10 min
-        duration: 10 * 1000, // for 10 secs
-        studyId: study.id))
-    ..addMeasure(NoiseMeasure(MeasureType(NameSpace.CARP, DataType.NOISE),
-        frequency: 10 * 60 * 1000, // sample sound every 10 min
-        duration: 10 * 1000, // for 10 secs
-        samplingRate: 500 // configure sampling rate to 500 ms
-        )));
-
-  study.addTask(SequentialTask('Task collecting a list of all installed apps')
+  study.addTask(Task('Task collecting a list of all installed apps')
     ..addMeasure(Measure(MeasureType(NameSpace.CARP, DataType.APPS))));
 
   // Create a Study Controller that can manage this study, initialize it, and start it.
   StudyController controller = StudyController(study);
-
-//  StudyManager manager =
-//  StudyManager(study, transformer: ((events) => events.where((event) => (event is BatteryDatum))));
-//
-
-//  StudyController controller = StudyController(
-//    study,
-//    transformer: ((events) => events.map((datum) {
-//          PrivacySchema.full().protect(datum);
-//        })),
-//    samplingSchema: SamplingSchema.common(),
-//  );
-
-  //manager = StudyManager(study, transformer: ((events) => events.transform(streamTransformer)));
-  controller.initialize();
+  await controller.initialize();
   controller.start();
 
   // listening on all data events from the study
   controller.events.forEach(print);
 
-  controller.events.where((datum) => datum.format.namepace == NameSpace.CARP);
+  // listen on only CARP events
+  controller.events.where((datum) => datum.format.namepace == NameSpace.CARP).forEach(print);
 
   // listening on a specific probe
   ProbeRegistry.probes[DataType.LOCATION].events.forEach(print);
@@ -83,11 +51,6 @@ void samplingSchemaExample() async {
           DataType.BLUETOOTH,
           PeriodicMeasure(MeasureType(NameSpace.CARP, DataType.BLUETOOTH),
               enabled: true, frequency: 60 * 60 * 1000, duration: 2 * 1000)),
-//      MapEntry(DataType.PHONE_LOG,
-//          PhoneLogMeasure(MeasureType(NameSpace.CARP, DataType.PHONE_LOG), enabled: true, days: 30)),
-      MapEntry(
-          DataType.TEXT_MESSAGE_LOG, Measure(MeasureType(NameSpace.CARP, DataType.TEXT_MESSAGE_LOG), enabled: true)),
-      MapEntry(DataType.TEXT_MESSAGE, Measure(MeasureType(NameSpace.CARP, DataType.TEXT_MESSAGE), enabled: true)),
     ]);
 
   // creating a sampling schema focused on activity and outdoor context (weather)
@@ -96,14 +59,6 @@ void samplingSchemaExample() async {
       MapEntry(DataType.PEDOMETER,
           PeriodicMeasure(MeasureType(NameSpace.CARP, DataType.PEDOMETER), enabled: true, frequency: 60 * 60 * 1000)),
       MapEntry(DataType.SCREEN, Measure(MeasureType(NameSpace.CARP, DataType.SCREEN), enabled: true)),
-      MapEntry(DataType.LOCATION, Measure(MeasureType(NameSpace.CARP, DataType.LOCATION), enabled: true)),
-      MapEntry(
-          DataType.NOISE,
-          PeriodicMeasure(MeasureType(NameSpace.CARP, DataType.NOISE),
-              enabled: true, frequency: 60 * 1000, duration: 2 * 1000)),
-      MapEntry(DataType.ACTIVITY, Measure(MeasureType(NameSpace.CARP, DataType.ACTIVITY), enabled: true)),
-//      MapEntry(DataType.WEATHER,
-//          WeatherMeasure(MeasureType(NameSpace.CARP, DataType.WEATHER), enabled: true, frequency: 2 * 60 * 60 * 1000))
     ]);
 
   //creating a study
@@ -148,7 +103,7 @@ void samplingSchemaExample() async {
 
   controller = StudyController(study);
   await controller.initialize();
-  await controller.start();
+  controller.start();
   print("Sensing started ...");
 
   // listening on all data events from the study