+++ carp_mobile_sensing/lib/probes/connectivity/connectivity_package.dart
@@ -34,7 +34,7 @@ class ConnectivitySamplingPackage implements SamplingPackage {
       MapEntry(
           DataType.BLUETOOTH,
           PeriodicMeasure(MeasureType(NameSpace.CARP, DataType.BLUETOOTH),
-              name: 'Nearby Devices (Bluetooth Scan)', enabled: true, frequency: 60 * 60 * 1000, duration: 2 * 1000)),
+              name: 'Nearby Devices (Bluetooth Scan)', enabled: true, frequency: 10 * 60 * 1000, duration: 2 * 1000)),
     ]);
 
   SamplingSchema get light => common