+++ carp_mobile_sensing/test/study_json_test.dart
@@ -7,7 +7,7 @@ String _encode(Object object) => const JsonEncoder.withIndent(' ').convert(objec
 
 void main() {
   setUp(() {
-    SamplingPackageRegistry.register(AudioSamplingPackage());
+    //SamplingPackageRegistry.register(AudioSamplingPackage());
     //SamplingPackageRegistry.register(CommunicationSamplingPackage());
     //SamplingPackageRegistry.register(ContextSamplingPackage());
   });